add superadmin config

+++ web-app/backend/controllers/userController.js
@@ -521,6 +521,16 @@ const deleteUser = asyncHandler(async (req, res) => {
   }
 
   try {
+    const user = await prisma.user.findUnique({ where: { id: userId } });
+
+    if (!user) {
+      return res.status(404).json({ message: "User not found" });
+    }
+
+    if (user.userRole === "superadmin") {
+      return res.status(403).json({ message: "Superadmin cannot be deleted" });
+    }
+
     await prisma.user.delete({ where: { id: userId } });
     res.json({ message: "User deleted successfully." });
   } catch (error) {
